Move OpacityControl to packages
