+++ src/packages/components/src/index.js
@@ -25,6 +25,7 @@ export { default as KadenceSelectTerms } from './select-terms-control';
 export { default as SmallResponsiveControl } from './small-responsive-control';
 export { default as ResponsiveControl } from './responsive-control';
 export { default as StepControls } from './step-control';
+export { default as OpacityControl } from './opacity-control';
 
 /* Links */
 export { default as DynamicLinkControl } from './links/dynamic-link-control';
